Add GitVersion fix for long dependabot branch name
Dependabot can create extremely long branch names, and the default
preview tag name approach in GitVersion incorporates this in the NuGet
version. This causes NuGet to produce warnings about over-long filenames
(because installing such a package in the local cache could cause
some filenames' lengths to exceed the default Win32 max path length.

So this defines a special branch type that detects Dependabot branches,
and picks a shorter name. We never actually release components from
these branches, so the preview tag name doesn't really matter, but we do
need it not to cause the NuGet warning, because we treat warning as
errors, so they fail the build.